+++ cx-portal/src/components/pages/AdminBoardDetail/components/BoardHeader/index.tsx
@@ -25,6 +25,7 @@ import { AppDetails } from 'features/apps/details/types'
 import { useFetchDocumentByIdMutation } from 'features/apps/apiSlice'
 import CommonService from 'services/CommonService'
 import './BoardHeader.scss'
+import { UseCaseType } from 'features/appManagement/types'
 
 export interface AppDetailHeaderProps {
   item: AppDetails
@@ -80,8 +81,8 @@ export default function BoardHeader({ item }: AppDetailHeaderProps) {
           <Typography variant="caption2" className="head">
             {t('content.adminboardDetail.usecase')}:{' '}
           </Typography>
-          {item.useCases.map((useCase) => (
-            <span key={useCase}> {useCase} </span>
+          {item.useCases.map((useCase: UseCaseType) => (
+            <span key={useCase.id}> {useCase.label} </span>
           ))}
         </div>
         <div className="price">